add test configs and files and restructure project directory
