+++ app_assetloader.py
@@ -1,0 +1,15 @@
+'''
+/*
+ * assetloader.py
+ *
+ * by joseyose, Oct 2021
+ *
+ * Provides the AssetLoader entry-point script.
+ *
+ */
+'''
+
+# from assetloader.app import main
+
+# if __name__ == "__main__":
+#     main()
