+++ requirements.txt
@@ -1,0 +1,1 @@
+pyqt5==5.15.4
