+++ setup.cfg
@@ -1,0 +1,41 @@
+[metadata]
+name = assetloader
+description = A GUI application for finding 3D assets and loading them into GPlay
+author = joseyose
+license = MIT
+license_file = LICENSE
+platforms = unix, linux, cyqwin, win32
+classifiers = 
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3 :: Only
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+
+[options]
+packages = 
+    assetloader
+install_requires = 
+    pyqt5>=5.15
+python_requires = >= 3.6
+package_dir = 
+    =src
+zip_safe = no
+
+[options.extras_require]
+testing = 
+    pytest>=6.0
+    pytest-cov>=2.0
+    mypy>=0.910
+    flake8>=3.9
+    tox>=3.24
+
+[options.package_data]
+assetloader = py.typed
+
+[flake8]
+max-line-length = 160
+# ignore = E128, E302, E305
+ignore = E711
+exclude = src\assetloader\ui\window.py src\assetloader\ui\resources\resources.py
