+++ src/assetloader/app.py
@@ -1,18 +1,18 @@
+import sys
+from PyQt5.QtWidgets import QApplication
+from .views import Window
 '''
 /*
  * app.py
- * 
+ *
  * by joseyose, Oct 2021
  *
- * Main driver for our GUI application. 
+ * Main driver for our GUI application.
  *
  */
 '''
 """This module provides the AssetLoader application."""
 
-import sys
-from PyQt5.QtWidgets import QApplication
-from .views import Window
 
 def main():
     app = QApplication(sys.argv)
