+++ src/assetloader/ui/resources/box.png
