+++ src/assetloader/ui/resources/resources.py
@@ -1,0 +1,136 @@
+# -*- coding: utf-8 -*-
+
+# Resource object code
+#
+# Created by: The Resource Compiler for PyQt5 (Qt v5.15.2)
+#
+# WARNING! All changes made in this file will be lost!
+
+from PyQt5 import QtCore
+
+qt_resource_data = b"\
+\x00\x00\x04\xf2\
+\x89\
+\x50\x4e\x47\x0d\x0a\x1a\x0a\x00\x00\x00\x0d\x49\x48\x44\x52\x00\
+\x00\x00\x18\x00\x00\x00\x18\x08\x06\x00\x00\x00\xe0\x77\x3d\xf8\
+\x00\x00\x00\x04\x73\x42\x49\x54\x08\x08\x08\x08\x7c\x08\x64\x88\
+\x00\x00\x00\x09\x70\x48\x59\x73\x00\x00\x00\xb1\x00\x00\x00\xb1\
+\x01\xc6\x2d\x49\x8d\x00\x00\x00\x19\x74\x45\x58\x74\x53\x6f\x66\
+\x74\x77\x61\x72\x65\x00\x77\x77\x77\x2e\x69\x6e\x6b\x73\x63\x61\
+\x70\x65\x2e\x6f\x72\x67\x9b\xee\x3c\x1a\x00\x00\x04\x6f\x49\x44\
+\x41\x54\x48\x89\xad\x55\x6b\x50\x54\x65\x18\x7e\xce\x39\x9c\x03\
+\x7b\x83\xe5\xb2\xec\xb2\xcb\x2e\xcb\xb2\x2b\x0e\xcb\x8a\x06\x66\
+\x3a\x64\x2a\x35\xa4\x4c\x8d\x9a\x30\xc3\x46\x65\x26\x58\x24\xe2\
+\x8c\xe1\x25\x8d\xc9\x7e\xd4\xa8\x31\x78\x19\x2d\x75\x2c\x0c\x2a\
+\x1c\x6d\xc3\xac\xc6\x1d\xc8\x99\x54\x24\x29\x47\x06\x98\x00\x2f\
+\x88\xe4\x0a\x23\xcb\x0a\xcb\x81\xbd\xef\xf6\xa3\x00\x17\x97\x8b\
+\x4d\xcf\xaf\x6f\xbe\x77\x9e\xe7\x79\x9f\xf7\xfb\xbe\x73\x80\x99\
+\x81\x08\xe3\x33\xd9\xf1\x72\x61\x73\xbc\x5c\xd8\x1c\xc6\x67\xb2\
+\x01\x10\x33\xe4\x4e\x09\x52\x22\xe2\x64\x6b\x94\x11\x6d\xdb\x0b\
+\x17\x0d\x3e\x6c\x29\xf1\x0d\xfd\xb9\xcd\xb7\xef\xfd\x8c\xa1\xd9\
+\xea\xa8\xdb\xd1\x51\xdc\x0d\x00\x82\xa6\x12\xa0\x26\xd9\xa7\x55\
+\x62\x61\x71\x5c\xa4\xf0\x5c\x92\x48\xf2\x9a\x54\x2a\x10\x6d\xca\
+\x9f\x1f\xac\x90\x85\x81\x61\x28\x2c\x4a\x93\x33\xeb\x72\xe6\x86\
+\x07\xd3\xd4\xd2\xae\x7b\x03\x6f\xd1\x34\xe5\x61\x47\x9c\xd7\x01\
+\x78\x27\x0a\x4d\x8c\x19\x22\x89\xe6\x6d\xf2\x38\xb1\x33\x55\x2c\
+\x13\x18\xd6\xe4\x11\x1c\x9a\xc6\xcd\x7e\x33\xf6\x5e\xfb\x15\xd6\
+\x10\x1b\x4a\x36\x2f\x44\x5a\x4a\xcc\x18\xc1\xe9\xf2\xe0\x8b\x53\
+\x4d\xf6\x43\x15\x8d\x66\x96\x75\x1d\xe9\xee\x19\x2c\x07\x60\x9f\
+\x98\x80\x17\x23\xe6\x6f\x15\x47\xf1\xbe\xdc\x5e\x98\x9e\x75\x6c\
+\x4f\x16\x7f\xd8\xeb\x24\x76\xd6\xd4\xc1\xe5\xf2\x60\x99\x52\x8d\
+\x95\x6a\x2d\x16\x08\x15\xf8\xdc\x70\x0d\xe5\x95\xbf\x41\x2a\x13\
+\x40\xa5\x10\x82\xa2\x48\xa4\xcd\x91\x06\xbd\x93\x97\x16\x2a\x13\
+\x0b\xd2\x5b\x6f\xf6\xad\x0f\x0a\x22\x64\xec\xb0\xeb\x2a\x00\x07\
+\x01\x40\xc3\xe7\xd2\xcd\xa5\x9b\x17\x07\x6d\xc9\x5f\x18\x44\x92\
+\xe3\xa1\xec\x0e\x37\x4e\x7c\xdd\x84\xef\x0c\xed\xc8\x55\xcf\xc5\
+\x1b\xc9\xa9\x60\x28\x0a\xbd\xec\x10\xca\xfe\xb8\x88\x36\xdb\x03\
+\x14\x6f\x9c\x8f\x17\x16\xab\xc6\x38\x5e\xaf\x0f\x65\xc7\x1b\xdc\
+\x1f\xed\xbf\xe8\x66\x47\x5c\x73\x08\x00\x4b\xd6\x66\x29\x2f\xd8\
+\x9c\x3e\x82\x17\xca\xc7\xb6\x8d\xcf\x61\x96\x2a\xd2\x6f\x6e\x4e\
+\x97\x07\x55\x67\x5a\x51\xf5\x6d\x0b\x56\xc6\x69\x91\xaf\x7b\x1a\
+\x1c\x9a\x46\xff\xc8\x08\x0e\x5c\xaf\x47\xa3\xa5\x1b\x85\x1b\x52\
+\x91\xa8\x89\xc4\xde\x23\x97\x31\xcc\xda\xc0\x61\x08\x5f\xc5\xd9\
+\x1b\xcb\x28\x00\x4a\x7d\xa6\xfc\xf5\xf2\x77\x93\x08\x8d\x84\xc1\
+\x27\x47\x7f\x87\xc1\xd8\x01\x8d\x4a\x04\x71\x14\xef\x9f\x39\x52\
+\x24\xe6\xe9\x24\xc8\xcb\xd1\xe1\x8e\xc3\x82\xad\xa7\xcf\xa3\xd7\
+\xca\xe2\x99\x18\x05\x5e\x8c\x9f\x05\x25\x27\x02\x39\xfb\x4e\xa1\
+\xad\xbd\x07\xa5\x05\xc9\x28\xcc\x99\x8d\x9b\x7f\x59\x7d\xb5\x0d\
+\xa6\xaf\xfc\xae\x58\x92\x52\x80\x93\x3b\x52\x70\xdb\x34\x8c\x3d\
+\x47\xea\x30\xe0\x20\x51\x52\xb8\x18\xf3\x53\xa4\xff\x1a\x11\xd0\
+\xaf\xd6\x22\x77\x95\x16\x35\xe7\x3b\xb0\xe6\x78\x25\x48\x1b\x01\
+\x11\x13\x82\x57\x67\x69\x91\xb1\x3c\x0a\xda\x84\x70\xbf\xf4\x64\
+\xa0\x3b\x9a\x20\xe3\xe1\xd8\x7b\x3a\x94\xe5\xab\x50\x59\x75\x09\
+\xfb\x0e\xfd\x02\xf8\xc6\xeb\x97\x1a\xbb\x71\xa1\xfe\x16\xce\x1e\
+\xcb\x00\x43\xfa\xb0\x27\x6d\x19\xe2\x04\x61\x81\xa4\xa6\x7e\x24\
+\xf2\x68\x0e\x0e\x16\x25\x01\x00\x8c\x35\x17\x71\xd0\xd0\x05\x90\
+\x14\x92\x55\xa1\xd8\xf5\xa6\x16\xf0\xb8\xa6\xa2\x4f\x6f\xf0\x28\
+\x32\x53\xc3\xd1\xd7\xcf\x02\x1c\x01\xf2\xb2\xd4\x00\x00\xbb\xc3\
+\x33\x56\xb7\xba\x9c\x20\x28\xfa\x31\x5e\xc0\x11\x4d\x06\x92\x20\
+\x40\x12\xfe\x6f\xb3\xfb\xa1\x15\x05\x57\x7f\x84\x39\xc6\x82\x25\
+\xe9\xe2\xc7\x38\x33\x4e\x10\x08\x0c\x4d\xe2\x95\x95\x0a\xac\x5b\
+\x95\x08\xa9\x88\x1b\xb8\xa9\xd1\xc5\x61\x43\x27\x8c\x8d\x0f\x9e\
+\xc8\x80\x24\x09\xec\x2a\x98\x37\xa9\xb8\x9f\x41\xee\xf3\xb1\x68\
+\x68\xb5\x20\x73\xcb\x15\x9c\xab\xef\x7d\x22\xa3\x29\x9b\x18\x5d\
+\x1c\x36\xdc\xc1\x7d\xb3\x1d\xfb\x8b\x74\xe8\xb5\xd8\xa7\xe2\x3c\
+\x11\xc6\xce\xe0\x83\xb5\x89\x68\xe9\xb4\x62\xc8\xe6\x46\xfe\x4b\
+\xca\xff\xd5\xc0\x54\x5d\x67\x62\x55\x12\x6e\xe8\xaa\x67\x25\x78\
+\xf4\x63\xf7\x5f\xe0\xf5\xfa\xf0\xfd\x85\xbb\xa8\xfe\xb9\x93\x05\
+\x60\xa2\x00\x58\x7a\xcc\xb6\x43\x6d\xdd\x76\xf2\xc4\x4f\x77\x75\
+\x04\x81\xe0\x14\x75\x18\x11\xc8\xa8\xa5\xd3\x0a\xd0\xc1\xd0\x69\
+\x22\x02\x0a\x9f\xa9\xed\xc2\xfa\xdd\x97\x59\xe3\x15\x53\x45\xf3\
+\x8d\xfe\x15\x00\xee\x4f\x54\x61\x9e\x4a\x8e\x2a\xa6\x81\x1d\xb9\
+\x19\x32\xe1\x86\x97\x15\x44\x08\x33\xfe\xd3\xfb\xa6\xf6\x1e\xc0\
+\x11\x40\xbf\x22\x61\x6c\xcf\xe9\xf2\xa2\xda\xd8\xe9\x3b\x7a\xba\
+\x9d\x1d\x71\xb8\x3f\x6b\x6a\x33\x7f\x08\xc0\x36\x5a\x9f\x6c\x1e\
+\x74\x42\x6c\xe8\xdb\x7c\x2e\x55\x9a\x9b\x21\x8b\x2c\x5a\x1d\x4f\
+\x70\x43\x28\x3f\x03\x87\xd3\x83\x93\x3f\xdc\xf2\x1d\x3d\xd3\x6e\
+\xb5\x0e\x3b\x0e\xdc\xea\xb6\x7e\x0c\xc0\x31\x51\x68\xba\x81\x93\
+\x49\xea\x70\x3d\x2f\x98\xfa\x74\xf9\x02\x51\xb4\x24\x9c\x21\x28\
+\xae\x00\x36\x87\xc7\x57\x6d\xbc\x33\xf0\x70\xd0\xb9\xbb\xa3\xcb\
+\x72\x18\x80\x7b\x1a\x9d\x69\x41\xc4\x8a\x39\x7a\x8d\x3c\xb4\x4f\
+\xa3\x08\xeb\x8b\x15\xf3\xf4\x33\x68\x0e\x00\xf0\x37\xb4\xe1\x82\
+\xb7\x45\x77\x4d\x87\x00\x00\x00\x00\x49\x45\x4e\x44\xae\x42\x60\
+\x82\
+"
+
+qt_resource_name = b"\
+\x00\x05\
+\x00\x6f\xa6\x53\
+\x00\x69\
+\x00\x63\x00\x6f\x00\x6e\x00\x73\
+\x00\x03\
+\x00\x00\x69\x68\
+\x00\x62\
+\x00\x6f\x00\x78\
+"
+
+qt_resource_struct_v1 = b"\
+\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x01\
+\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x02\
+\x00\x00\x00\x10\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\
+"
+
+qt_resource_struct_v2 = b"\
+\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x01\
+\x00\x00\x00\x00\x00\x00\x00\x00\
+\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x02\
+\x00\x00\x00\x00\x00\x00\x00\x00\
+\x00\x00\x00\x10\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\
+\x00\x00\x01\x7c\x3e\xc0\xdd\x3c\
+"
+
+qt_version = [int(v) for v in QtCore.qVersion().split('.')]
+if qt_version < [5, 8, 0]:
+    rcc_version = 1
+    qt_resource_struct = qt_resource_struct_v1
+else:
+    rcc_version = 2
+    qt_resource_struct = qt_resource_struct_v2
+
+def qInitResources():
+    QtCore.qRegisterResourceData(rcc_version, qt_resource_struct, qt_resource_name, qt_resource_data)
+
+def qCleanupResources():
+    QtCore.qUnregisterResourceData(rcc_version, qt_resource_struct, qt_resource_name, qt_resource_data)
+
+qInitResources()
