+++ tox.ini
@@ -1,0 +1,30 @@
+[tox]
+minversion = 3.9.0
+envlist = py36, py37, py38, py39, flake8, mypy
+
+[gh-actions]
+python 
+    3.6: py36
+    3.7: py37
+    3.8: py38
+    3.9: py39, mypy, flake8
+
+[testenv]
+setenv = 
+    PYTHONPATH = {toxinidir}
+deps = 
+    -r{toxinidir}/requirements_dev.txt
+commands = 
+    pytest --basetemp={envtmpdir}
+
+[testenv:flake8]
+basepython = python3.9
+deps = flake8
+commands = flake8 src tests
+
+[testenv:mypy]
+basepython = python3.9
+deps = 
+    -r{toxinidir}/requirements_dev.txt
+command = mypy src
+
